The full backend system flow for orders to delivery is complete
An Order can be created.
The order is split into SubOrders.
SubOrders can be marked as ready for collection.
The system checks if all SubOrders of a given Order are ready.
If they are, they are marked for collection and the driver can see the new order.

Note; This is an entirely unidirectional implementation. It would likely be more efficient and cleaner to use a bidirectional system.