Added functionality to deliveries and fulfilments
Added ability to mark orders as delivered in the deliveries page.

Both deliveries and fulfilment pages refresh when marked as delivered/ready. (note: this would be better done without refreshing the page)

A confirmation dialog box is opened when attempting to clear a delivery or suborder.

Other changes:
ManyToMany relationship added for suppliers and orders. Some work might be needed to clean the join table of any orders that have been cleared.