Lots of changes
Most notably:

Messages are given when a user is not logged in and tries to view certain pages or views products.

Product price calculations are working, cart price is half working.

Navbar will now react to the user role (need to finish implementing, would be ideal to use Thymeleaf layout but it's not working for some reason so might just have to add the navbar manually to every template/page)