Many changes, mostly to the product pages.
The product page now uses clip-path to display the image as a circle.

Added a modal to the product page for image.

Name of the Supplier is now included on the product page.

If an item is already in a customer's cart they are shown a message and link to their cart instead of allowing "Add To Cart".

Prices now have double digits after the decimal point.