Started implementing error handling for some login errors. Fulfilment page is now better. Stock is now reduced upon checkout.
Upon checkout, item stocks are reduced by the quantity in the order. Users can only "add to cart" what's available and are informed if an item is out of stock and won't be able to "add to cart". There exists an exception if the user is looking at an item while someone else checks out with it, so additional code is being added to the controller to cancel the order and inform the user.

Started work on preventing whitepage error messages when a user isn't logged in and tries to do something they can't.

Also redesigned the fulfilments page to make it look nicer and show more information, as well as just be more useful in general.