Many Improvements, mainly to checkout process.
Checkout page now prevents users from checkout if an item quantity exceeds the product stock. This value is updated when the user visits the checkout or changes the quantity in their cart. That alone wouldn't prevent an invalid order, so code was added to the ShoppingCartController class to handle that too. The user will be given a message that their order couldn't be placed and why, and a button will return them to their cart.

Fixed a styling issue where the product card images were all different heights.

Made general improvements to the appearance of the checkout page.

Location choices on registration page are now dynamic rather than hardcoded.