Added Google Maps API for Delivery Drivers
Right now it just points from Telemark to Oslo in Norway, so it might be prudent to create some real addresses for testing this properly. The origin could always be the same since we don't have real drivers, but in the real world the origin would be their vehicle's location.