Added products to front page, fixed navbar issues.
Navbar used "userRole" to determine the fragment to use, but it was conflicting with other code, so I changed it to use its own variable "navbarRole" which is only updated when the navbar is.

Created a new front page that's more inline with the original wireframes. Not 100% sold on the solution so I might rework it.